Add styles (less) for page rendering endpoint. Re GH-71
