+++ client/colors.less
@@ -1,0 +1,14 @@
+@blue: #669eff;
+@shadowBlue: #5793f2;
+@sapphire: #315d8a;
+@teal: #61d3c9;
+@slate: #6d7886;
+@heatherGrey: #cdd2dd;
+@lightGrey: #e9ebf0;
+@softGrey: #fafcff;
+@plum: #404467;
+@shadowPlum: #31344f;
+@brick: #EB5D52;
+@yellow: #F0CF62;
+@white: #fff;
+@black: #000;
