+++ client/main.less
@@ -1,0 +1,34 @@
+@import "normalize";
+@import "colors";
+
+// Components -----------------------------------------------------------------
+@import "components/element-group/element-group";
+@import "components/el/el";
+
+// Pages ----------------------------------------------------------------------
+@import "pages/page/page";
+
+// General --------------------------------------------------------------------
+*,
+*:after,
+*:before {
+  box-sizing: border-box;
+  -webkit-backface-visibility: hidden;
+  -webkit-tap-highlight-color: rgba(0,0,0,0);
+  -webkit-tap-highlight-color: transparent;
+}
+
+html {
+  height: 100%;
+}
+
+body {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+  font-family: "FiraSans", "Fira Sans", "Fira Sans OT", sans-serif;
+}
+
+#app {
+  height: 100%;
+}
