+++ client/pages/page/page.less
@@ -1,0 +1,30 @@
+#page {
+  @_gutter: 10px;
+  @_baseWidth: 320px;
+  @_baseHeight: 440px;
+
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: @lightGrey;
+
+  .container {
+    position: relative;
+    width: @_baseWidth;
+    height: @_baseHeight;
+
+    .inner {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      background: @softGrey;
+    }
+  }
+}
