+++ package.json
@@ -8,6 +8,7 @@
     "build": "npm-run-all build:**",
     "build:clean": "rimraf ./build && mkdirp ./build",
     "build:static": "ncp ./static ./build",
+    "build:css": "lessc ./client/main.less ./build/css/main.bundle.css",
     "build:js": "webpack",
     "test": "npm-run-all test:**",
     "test:jshint": "jshint -c node_modules/mofo-style/linters/.jshintrc server static/js server.js",
@@ -33,6 +34,7 @@
     "xhr": "2.0.1"
   },
   "devDependencies": {
+    "autoless": "~0.1.7",
     "babel": "~5.4.7",
     "babel-core": "~5.4.7",
     "babel-loader": "~5.1.3",
