+++ static/render.html
@@ -8,8 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Webmaker</title>
 
-    <link rel="stylesheet" href="./css/normalize.css">
-    <link rel="stylesheet" href="./css/main.css">
+    <link rel="stylesheet" href="./css/main.bundle.css">
     <link rel="stylesheet" href="//code.cdn.mozilla.net/fonts/fira.css">
   </head>
